Removes wrapper from upgrade.
To test a possible timeout with Jenkins and only have raw
output provided in logs, upgrade function is changes to
run the upgrade in just bash. This will result in the
upgrade being determined to be successful or not based only
on the return script of the upgrade, rather than parsing for
failures.

Change-Id: I183cc86d4a19988e337cbfd037ebaa806e3865ef